simple_version/requirements.txt
# Web automation
playwright==1.47.0

# OpenAI API
openai==1.52.0

# Anthropic API for Claude
anthropic==0.3.11

# Google AI for Gemini
google-generativeai==0.3.1

# Async support
asyncio==3.4.3

# # Backoff for API rate limiting
# backoff==2.2.1

# Transformers for tokenizer
transformers==4.36.2

python-dotenv==1.0.1
# Development tools
pytest==7.4.3
black==23.12.1
isort==5.13.2
flake8==6.1.0

tenacity=9.0.0
